differentiated student and teacher pages
